requirements.txt
# Telegram Bot API (비동기 지원 최신 버전)
python-telegram-bot>=21.0

# Google Gemini API
google-generativeai>=0.8.0

# 환경변수 관리
python-dotenv>=1.0.1

# 백그라운드 스케줄러 (먼저 말 걸기 기능 및 야간 리포트)
APScheduler>=3.11.0

# 타임존 처리 (APScheduler 필수 의존성)
pytz

# 비동기 HTTP 요청 (날씨 API 호출용)
httpx
